+++ pages/index.js
@@ -1,1 +1,4 @@
-export default () => <div>Welcome to next.</div>;
+import App from '../src/App';
+import { loginForm } from '../config/forms';
+
+export default () => <App form={loginForm} />;
