+++ src/App.js
@@ -1,57 +1,22 @@
 import React, { Component } from 'react';
-import {
-  MuiThemeProvider,
-  createMuiTheme,
-  withStyles,
-  CssBaseline,
-  Grid,
-  colors
-} from '@material-ui/core';
+import { Grid } from '@material-ui/core';
 import Card from './components/Card';
 import Decoration from './components/Decoration';
 import FormContainer from './components/FormContainer';
-import { loginForm } from './config/forms';
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: colors.lightBlue[100],
-      main: colors.lightBlue[200],
-      dark: colors.lightBlue[300],
-      contrastText: '#fff'
-    }
-  }
-});
-
-const styles = {
-  app: {
-    background: `radial-gradient(${colors.lightBlue[50]}, ${
-      colors.lightBlue[300]
-    })`,
-    minHeight: '100vh'
-  }
-};
+import MaterialContext from './components/MaterialContext';
 
 class App extends Component {
   render() {
-    const { classes } = this.props;
+    const { form } = this.props;
     return (
-      <MuiThemeProvider theme={theme}>
-        <CssBaseline />
-        <Grid
-          container
-          justify="center"
-          alignItems="center"
-          className={classes.app}
-        >
-          <Card>
-            <Decoration />
-            <FormContainer form={loginForm} />
-          </Card>
-        </Grid>
-      </MuiThemeProvider>
+      <MaterialContext>
+        <Card>
+          <Decoration />
+          <FormContainer form={form} />
+        </Card>
+      </MaterialContext>
     );
   }
 }
 
-export default withStyles(styles)(App);
+export default App;
